Lesson_39/requirements.txt
aiohttp
bcrypt
PyJWT
python-dotenv
